Add CountyCalibrations to the admin and revamp the admin index page with a simple menu.
